UBCx/4. Object-Oriented Design/robust-classes-practice-problems/FoodTruck/src/model/Owner.java
package model;

import model.exceptions.NoCookException;
import model.exceptions.NoIngredientException;
import model.exceptions.NotEnoughMoneyException;
import org.omg.CosNaming.NamingContextPackage.NotEmpty;

public class Owner {

    private String name;
    private Kitchen kitchen;

    public Owner(String name, Kitchen kit) {
        this.name = name;
        kitchen = kit;
    }

    //getters
    public String getName() { return name; }
    public Kitchen getKitchen() { return kitchen; }

    // REQUIRES: we have enough ingredients to make the specified amount of tacos
    // MODIFIES: this
    // EFFECTS: calls makeTaco on the kitchen, returns true
    public boolean orderMoreTacos(int amount) throws NoCookException, NotEnoughMoneyException {
        kitchen.makeTaco(amount);
        return true;
    }

    //MODIFIES: this 
    //EFFECTS: calls buyIngredients on the kitchen, returns true.
    public boolean askForMoreIngredients(int amount) throws NotEnoughMoneyException {
        this.kitchen.buyIngredients(amount);
        return true;
    }


}
